refactor: minor code refactor
